:recycle: Split ga logic
